--- examples/data/custom_encoded.csv
@@ -1,10 +1,0 @@
-V1,V2,V3,V4,V5,V6,V7,V8,V9,V10
-0             ,12    ,9        ,2                ,5              ,2    ,3     ,3       ,3           ,2
-2             ,7     ,11       ,6                ,3              ,3    ,3     ,0       ,1           ,4
-5             ,15    ,15       ,11               ,2              ,1    ,1     ,3       ,3           ,4
-5             ,15    ,11       ,11               ,2              ,3    ,4     ,3       ,3           ,4
-0             ,7     ,9        ,7                ,5              ,2    ,2     ,3       ,2           ,1
-3             ,7     ,9        ,5                ,5              ,0    ,0     ,1       ,1           ,3
-0             ,4     ,9        ,8                ,5              ,2    ,2     ,0       ,1           ,2
-0             ,4     ,9        ,8                ,5              ,3    ,2     ,0       ,1           ,3
-0             ,12    ,11       ,2                ,5              ,0    ,1     ,3       ,1           ,2
