+++ test/unit/pipelines/test_pipeline_ts_wrappers.py
@@ -4,7 +4,6 @@ import numpy as np
 import pandas as pd
 from sklearn.metrics import mean_absolute_percentage_error
 
-from examples.advanced.time_series_forecasting.composing_pipelines import visualise
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
 from fedot.core.pipelines.node import PipelineNode
